--- src/components/index.ts
@@ -1,1 +1,0 @@
-// export * from "./Button"; 형식으로
